chore: remove unused old storage engien and other components
