+++ Cargo.toml
@@ -2,13 +2,10 @@
 
 members = [
     "foyer",
-    "foyer-bench",
     "foyer-common",
     "foyer-intrusive",
-    "foyer-policy",
     "foyer-storage",
     "foyer-storage-bench",
-    "foyer-utils",
 ]
 
 [patch.crates-io]
