+++ foyer-common/Cargo.toml
@@ -9,3 +9,5 @@ license = "Apache-2.0"
 
 [dependencies]
 bytes = "1"
+paste = "1.0"
+tokio = { version = "1", features = ["sync"] }
