+++ foyer-common/src/bits.rs
