--- foyer-policy/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "foyer-policy"
-version = "0.1.0"
-edition = "2021"
-authors = ["MrCroxx <mrcroxx@outlook.com>"]
-description = "Hybrid cache for Rust"
-license = "Apache-2.0"
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-bytes = "1"
-cmsketch = "0.1"
-foyer-utils = { path = "../foyer-utils" }
-itertools = "0.10.5"
-memoffset = "0.8"
-parking_lot = "0.12"
-paste = "1.0"
-prometheus = "0.13"
-rand = "0.8.5"
-thiserror = "1"
-tokio = { version = "1", features = [
-    "rt",
-    "rt-multi-thread",
-    "sync",
-    "macros",
-    "time",
-    "signal",
-] }
-tracing = "0.1"
-twox-hash = "1"
-
-
