--- foyer-policy/src/lib.rs
@@ -1,30 +1,0 @@
-//  Copyright 2023 MrCroxx
-//
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//  http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-
-#![feature(trait_alias)]
-#![feature(pattern)]
-
-pub mod eviction;
-pub mod reinsertion;
-
-pub trait Item = PartialOrd
-    + Ord
-    + PartialEq
-    + Eq
-    + Clone
-    + std::hash::Hash
-    + Send
-    + Sync
-    + 'static
-    + std::fmt::Debug;
