+++ foyer-storage/Cargo.toml
@@ -14,7 +14,6 @@ bytes = "1"
 cmsketch = "0.1"
 foyer-common = { path = "../foyer-common" }
 foyer-intrusive = { path = "../foyer-intrusive" }
-foyer-utils = { path = "../foyer-utils" }
 futures = "0.3"
 itertools = "0.10.5"
 libc = "0.2"
