--- foyer-utils/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "foyer-utils"
-version = "0.1.0"
-edition = "2021"
-authors = ["MrCroxx <mrcroxx@outlook.com>"]
-description = "Hybrid cache for Rust"
-license = "Apache-2.0"
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-bytes = "1"
-itertools = "0.10.5"
-memoffset = "0.8"
-parking_lot = "0.12"
-paste = "1.0"
-thiserror = "1"
-tokio = { version = "1", features = [
-    "rt",
-    "rt-multi-thread",
-    "sync",
-    "macros",
-    "time",
-    "signal",
-] }
-
