+++ foyer/Cargo.toml
@@ -13,8 +13,6 @@ bytes = "1"
 cmsketch = "0.1"
 crossbeam = "0.8"
 foyer-common = { path = "../foyer-common" }
-foyer-policy = { path = "../foyer-policy" }
-foyer-utils = { path = "../foyer-utils" }
 futures = "0.3"
 itertools = "0.10.5"
 libc = "0.2"
